refactor: chat 컴포넌트 header, main, footer로 모듈화
